requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.1
langchain==0.3.7
langchain-google-genai==2.0.3
google-generativeai==0.7.2